修复数据类型传递错误问题

+++ modules/iopcore/cdvdman/cdvdman.c
@@ -205,7 +205,7 @@ int read_raw_data(u8 *addr, u32 size, u32 offset, u32 shift)
     // read first block if not aligned to sector size
     if (pos) {
         int r = MIN(size, (2048 - pos));
-        ReadSectors(lba, ziso_tmp_buf, 1);
+        ReadSectors((u64)lba, ziso_tmp_buf, 1);
         memcpy(addr, ziso_tmp_buf + pos, r);
         size -= r;
         lba++;
@@ -218,7 +218,7 @@ int read_raw_data(u8 *addr, u32 size, u32 offset, u32 shift)
         n_blocks++;
     if (n_blocks > 1) {
         int r = 2048 * (n_blocks - 1);
-        ReadSectors(lba, addr, n_blocks - 1);
+        ReadSectors((u64)lba, addr, n_blocks - 1);
         size -= r;
         addr += r;
         lba += n_blocks - 1;
@@ -226,7 +226,7 @@ int read_raw_data(u8 *addr, u32 size, u32 offset, u32 shift)
 
     // read remaining data
     if (size) {
-        ReadSectors(lba, ziso_tmp_buf, 1);
+        ReadSectors((u64)lba, ziso_tmp_buf, 1);
         memcpy(addr, ziso_tmp_buf, size);
         size = 0;
     }
